+++ packages/cloud_firestore/cloud_firestore/example/test_driver/snapshot_metadata_e2e.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// @dart=2.9
-
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
@@ -14,7 +12,7 @@ void runSnapshotMetadataTests() {
   group(
     '$SnapshotMetadata',
     () {
-      FirebaseFirestore /*?*/ firestore;
+      late FirebaseFirestore /*?*/ firestore;
 
       setUpAll(() async {
         firestore = FirebaseFirestore.instance;