+++ packages/cloud_functions/cloud_functions/example/test_driver/cloud_functions_e2e.dart
@@ -1,17 +1,13 @@
-// ignore_for_file: require_trailing_commas
-// @dart = 2.9
-
 // Copyright 2020, the Chromium project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-
-// @dart=2.9
 
 import 'package:cloud_functions/cloud_functions.dart';
 import 'package:drive/drive.dart' as drive;
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'firebase_config.dart';
 import 'sample.dart' as data;
 
 String kTestFunctionDefaultRegion = 'testFunctionDefaultRegion';
@@ -20,9 +16,10 @@ String kTestFunctionTimeout = 'testFunctionTimeout';
 String kTestMapConvertType = 'testMapConvertType';
 
 void testsMain() {
-  HttpsCallable callable;
+  late HttpsCallable callable;
+
   setUpAll(() async {
-    await Firebase.initializeApp();
+    await Firebase.initializeApp(options: TestFirebaseConfig.platformOptions);
     FirebaseFunctions.instance.useFunctionsEmulator('localhost', 5001);
     callable =
         FirebaseFunctions.instance.httpsCallable(kTestFunctionDefaultRegion);
@@ -85,15 +82,16 @@ void testsMain() {
     });
 
     test(
-        '[HttpsCallableResult.data] should return Map<String, dynamic> type for returned objects',
-        () async {
-      HttpsCallable callable =
-          FirebaseFunctions.instance.httpsCallable(kTestMapConvertType);
+      '[HttpsCallableResult.data] should return Map<String, dynamic> type for returned objects',
+      () async {
+        HttpsCallable callable =
+            FirebaseFunctions.instance.httpsCallable(kTestMapConvertType);
 
-      var result = await callable();
+        var result = await callable();
 
-      expect(result.data, isA<Map<String, dynamic>>());
-    });
+        expect(result.data, isA<Map<String, dynamic>>());
+      },
+    );
   });
 
   group('FirebaseFunctionsException', () {
@@ -107,7 +105,7 @@ void testsMain() {
         expect(e.message, equals('Invalid test requested.'));
         return;
       } catch (e) {
-        fail(e);
+        fail('$e');
       }
     });
 
@@ -122,18 +120,21 @@ void testsMain() {
       } on FirebaseFunctionsException catch (e) {
         expect(e.code, equals('cancelled'));
         expect(
-            e.message,
-            equals(
-                'Response data was requested to be sent as part of an Error payload, so here we are!'));
+          e.message,
+          equals(
+            'Response data was requested to be sent as part of an Error payload, so here we are!',
+          ),
+        );
         expect(e.details, equals(data.deepMap));
       } catch (e) {
-        fail(e);
+        fail('$e');
       }
     });
   });
 
   group('region', () {
-    HttpsCallable customRegionCallable;
+    late HttpsCallable customRegionCallable;
+
     setUpAll(() async {
       customRegionCallable =
           FirebaseFunctions.instanceFor(region: 'europe-west1')
@@ -147,12 +148,13 @@ void testsMain() {
   });
 
   group('HttpsCallableOptions', () {
-    HttpsCallable timeoutCallable;
+    late HttpsCallable timeoutCallable;
 
     setUpAll(() async {
       timeoutCallable = FirebaseFunctions.instance.httpsCallable(
-          kTestFunctionTimeout,
-          options: HttpsCallableOptions(timeout: const Duration(seconds: 3)));
+        kTestFunctionTimeout,
+        options: HttpsCallableOptions(timeout: const Duration(seconds: 3)),
+      );
     });
 
     test('times out when the provided timeout is exceeded', () async {
@@ -164,7 +166,7 @@ void testsMain() {
       } on FirebaseFunctionsException catch (e) {
         expect(e.code, equals('deadline-exceeded'));
       } catch (e) {
-        fail(e);
+        fail('$e');
       }
     });
   });