+++ packages/firebase_database/firebase_database/ios/Classes/FLTFirebaseDatabasePlugin.h
@@ -1,4 +1,4 @@
-// Copyright 2020 The Chromium Authors. All rights reserved.
+// Copyright 2021 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 #import <Firebase/Firebase.h>
@@ -8,10 +8,8 @@
 #import <Flutter/Flutter.h>
 #endif
 
-@interface FLTFirebaseDatabasePlugin : NSObject <FlutterPlugin>
+#import <Foundation/Foundation.h>
+#import <firebase_core/FLTFirebasePlugin.h>
 
-+ (NSMutableArray *)getSnapshotChildKeys:(FIRDataSnapshot *)dataSnapshot;
-
-@property(nonatomic) NSMutableDictionary *updatedSnapshots;
-
+@interface FLTFirebaseDatabasePlugin : FLTFirebasePlugin <FlutterPlugin, FLTFirebasePlugin>
 @end