+++ website/dictionary.js
@@ -45,6 +45,7 @@ module.exports = [
   'filesystem',
   'firebase',
   'firestore',
+  'flutterfire',
   'FlutterFire',
   'func',
   'getter',