+++ packages/firebase_app_check/firebase_app_check_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3+1
+
+ - Update a dependency to the latest release.
+
+## 0.0.3
+
+ - **FEAT**: automatically inject Firebase JS SDKs (#7359).
+
 ## 0.0.2
 
  - **FEAT**: support initializing default `FirebaseApp` instances from Dart (#6549).