+++ packages/firebase_core/firebase_core_web/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.4.0
+
+ - **FEAT**: bump Firebase JS SDK to `8.10.0` (#7460).
+
+## 1.3.0
+
+ - **FEAT**: automatically inject Firebase JS SDKs (#7359).
+ - **FEAT**: auto inject Firebase scripts (#7358).
+
 ## 1.2.0
 
  - **FEAT**: support initializing default `FirebaseApp` instances from Dart (#6549).