+++ packages/firebase_messaging/firebase_messaging/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 11.2.0
+
+ - **REFACTOR**: migrate remaining examples & e2e tests to null-safety (#7393).
+ - **FEAT**: automatically inject Firebase JS SDKs (#7359).
+
 ## 11.1.0
 
  - **FEAT**: support initializing default `FirebaseApp` instances from Dart (#6549).